ocr/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pytesseract==0.3.10
Pillow==10.3.0
python-multipart>=0.0.9
pdf2image==1.17.0
python-magic==0.4.27
